requirements.txt
dependency-injector==4.42.0
watchdog==5.0.3
plex-api-client==0.18.1
python-telegram-bot==21.6
celery==5.4.0
redis==5.1.1